+++ lib/core/style/c_colors.dart
@@ -1,0 +1,11 @@
+import 'dart:ui';
+
+class CColors {
+  static const Color white = Color(0xffffffff);
+  static const Color accentSoft = Color(0xffF5FFF9);
+  static const Color accent = Color(0xffE1F3E2);
+  static const Color primary = Color(0xffC8E1C9);
+  static const Color transparentPink = Color(0x30ff69a9);
+  static const Color textDark = Color(0xff51657c);
+  static const Color text = Color(0xff2E476E);
+}
