--- lib/data/datasource/datasource.dart
@@ -1,5 +1,0 @@
-class Datasource {
-
-  // TODO: implement work with Firebase
-
-}
