--- lib/data/models/card_model.dart
@@ -1,5 +1,0 @@
-import '../../domain/entities/card.dart';
-
-class CardModel extends Card {
-  // TODO: implement flash card's model
-}
