+++ lib/features/cards/data/models/card_model.dart
@@ -1,0 +1,5 @@
+import '../../domain/entities/card.dart';
+
+class CardModel extends Card {
+  // TODO: implement flash card's model
+}
