+++ lib/features/cards/domain/entities/card.dart
