+++ lib/features/profile/presentation/pages/user_profile_page.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class UserProfilePage extends StatelessWidget {
+  const UserProfilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        children: [
+          Container(
+            padding: EdgeInsets.all(70),
+            alignment: Alignment.center,
+            child: Text(
+              "User Profile Page",
+              style: TextStyle(fontSize: 25),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
