--- lib/presentation/pages/main_page.dart
@@ -1,74 +1,0 @@
-import 'package:flutter/material.dart';
-import '/presentation/pages/homepage.dart';
-import '/presentation/pages/study_materials_page.dart';
-import 'to_do/todo_page.dart';
-import '/presentation/pages/user_profile_page.dart';
-import '/presentation/widgets/bottom_navigation_bar.dart';
-
-import '../../data/datasource/c_routes.dart';
-
-class MainPage extends StatelessWidget {
-  final _navigatorKey = GlobalKey<NavigatorState>();
-
-  MainPage({super.key});
-
-  Route _onGenerateRoute(RouteSettings settings) {
-    late Widget page;
-
-    switch (settings.name) {
-      case CRoutes.routeHomepage:
-        page = HomePage();
-      case CRoutes.routeStudyMaterials:
-        page = StudyMaterialsPage();
-      case CRoutes.routeToDo:
-        page = ToDoPage();
-      case CRoutes.routeUserProfile:
-        page = UserProfilePage();
-      default:
-        page = const Placeholder();
-    }
-
-    return PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => page,
-      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-        const begin = Offset(0.0, 1.0);
-        const end = Offset.zero;
-        const curve = Curves.ease;
-
-        var tween =
-            Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
-
-        return SlideTransition(
-          position: animation.drive(tween),
-          child: child,
-        );
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return PopScope(
-      canPop: false,
-      child: Stack(
-        fit: StackFit.expand,
-        children: [
-          // page itself
-          Navigator(
-            key: _navigatorKey,
-            initialRoute: CRoutes.routeHomepage,
-            onGenerateRoute: _onGenerateRoute,
-          ),
-
-          // template with a custom bottom navigation bar
-          Column(
-            children: [
-              const Expanded(child: SizedBox()),
-              CustomBottomNavBar(_navigatorKey),
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
